+++ pyengine/Utils/Font.py
@@ -1,58 +1,98 @@
+from pyengine.Utils.Color import Color, Colors
+
 import pygame
 
 __all__ = ["Font"]
 
 
 class Font:
-    def __init__(self, name: str = "Arial", size: int = 15, bold: bool = False, italic: bool = False):
+    def __init__(self, name: str = "arial", size: int = 15, bold: bool = False, italic: bool = False,
+                 underline: bool = False, color: Color = Colors.WHITE.value, background: Color = None,
+                 antialias: bool = True):
+        pygame.font.init()
+        try:
+            self.font = pygame.font.Font(name, size)
+        except FileNotFoundError:
+            self.font = pygame.font.SysFont(name, size)
         self.name = name
         self.size = size
         self.bold = bold
         self.italic = italic
+        self.underline = underline
+        self.color = color
+        self.background = background
+        self.antialias = antialias
         self.parent = None
 
-    @property
-    def name(self):
-        return self.__name
+    def get_ascent(self):
+        return self.font.get_ascent()
 
-    @name.setter
-    def name(self, name):
-        self.__name = name
+    def get_bold(self):
+        return self.font.get_bold()
 
-    @property
-    def size(self):
-        return self.__size
+    def get_descent(self):
+        return self.font.get_descent()
 
-    @size.setter
-    def size(self, size):
-        self.__size = size
+    def get_height(self):
+        return self.font.get_height()
 
-    @property
-    def bold(self):
-        return self.__bold
+    def get_italic(self):
+        return self.font.get_italic()
 
-    @bold.setter
-    def bold(self, bold):
-        self.__bold = bold
+    def get_linesize(self):
+        return self.font.get_linesize()
 
-    @property
-    def italic(self):
-        return self.__italic
+    def get_underline(self):
+        return self.font.get_underline()
 
-    @italic.setter
-    def italic(self, italic):
-        self.__italic = italic
+    def get_antialias(self):
+        return self.__antialias
 
-    def __eq__(self, other) -> bool:
-        return self.name == other.name and self.size == other.size and\
-               self.bold == other.bold and self.italic == other.italic
+    def get_color(self):
+        return self.__color
 
-    def __repr__(self) -> str:
-        return str([self.name, self.size, self.bold, self.italic])
+    def get_background(self):
+        return self.__background
 
-    def rendered_size(self, texte):
-        rect = self.render().render(texte, 1, (0, 0, 0)).get_rect()
-        return rect.width, rect.height
+    def metrics(self, text):
+        return self.font.get_metrics(text)
 
-    def render(self) -> pygame.font:
-        return pygame.font.SysFont(self.name, self.size, self.bold, self.italic)
+    def render(self, text):
+        return self.font.render(text, self.antialias, self.color.get(), self.background)
+
+    def set_bold(self, boldbool):
+        self.font.set_bold(boldbool)
+
+    def set_italic(self, italicbool):
+        self.font.set_italic(italicbool)
+
+    def set_underline(self, underlinebool):
+        self.font.set_underline(underlinebool)
+
+    def set_antialias(self, antialiasebool):
+        self.__antialias = antialiasebool
+
+    def set_color(self, color):
+        if not isinstance(color, Color):
+            raise TypeError("Color must be a color")
+        self.__color = color
+
+    def set_background(self, background):
+        if not isinstance(background, Color) and background is not None:
+            raise TypeError("Background must be a color")
+        self.__background = background
+
+    def rendered_size(self, text):
+        return self.font.size(text)
+
+    bold = property(get_bold, set_bold)
+    italic = property(get_italic, set_italic)
+    underline = property(get_underline, set_underline)
+    antialias = property(get_antialias, set_antialias)
+    color = property(get_color, set_color)
+    background = property(get_background, set_background)
+
+    def __eq__(self, other):
+        return (isinstance(other, Font) and self.name == other.name and self.color == other.color and
+                self.background == other.background and self.bold == other.bold and self.italic == other.italic and
+                self.underline == other.underline and self.antialias == other.antialias)